+++ packages/isar/lib/src/common/isar_link_base_impl.dart
@@ -83,6 +83,7 @@ abstract class IsarLinkBaseImpl<OBJ> implements IsarLinkBase<OBJ> {
       collection: targetCollection,
       whereClauses: [
         LinkWhereClause(
+          linkCollection: sourceCollection.name,
           linkName: linkName,
           id: containingId,
         ),