+++ packages/isar_generator/lib/src/code_gen/query_filter_generator.dart
@@ -1,7 +1,6 @@
 import 'package:dartx/dartx.dart';
 import 'package:isar/isar.dart';
 import 'package:isar_generator/src/code_gen/query_filter_length.dart';
-import 'package:isar_generator/src/isar_type.dart';
 import 'package:isar_generator/src/object_info.dart';
 
 class FilterGenerator {